Bugfix Command Search
Fixed it so user can search for commands while command list gets filtered and shows. some bugs with this that needs to be fixed. 1 cant click the dropdown away, 2 dropdown stays still when window is dragged, 3 why can i not add custom plugins? must be because it updates from commands.json...